requirements.txt
streamlit==1.38.0
Pillow==10.4.0
reportlab==4.2.2
pillow-heif==0.18.0